fix(ui): resolve 6 Kotlin compile errors from CI log review
Pulled run #3 log ZIP (auth-strip redirect) and fixed each 'e:' line:
1. PositionRing.kt:70 - add androidx.compose.runtime.getValue import (infinite.animateFloat delegate)
2. PriceSparkline.kt:56 - norm() y math returned Double -> toFloat() for Pair<Float,Float>
3. DashboardScreen.kt:111 - stray bare Spacer() inside LazyColumn LazyListScope (not @Composable ctx) -> removed
4. DashboardScreen.kt:287 - missing import com.titanium.trades.ui.components.PositionRing
5-6. SettingsScreen.kt:156,186 - missing androidx.compose.ui.unit.sp import for letterSpacing

+++ app/src/main/java/com/titanium/trades/ui/components/PriceSparkline.kt
@@ -50,7 +50,7 @@ fun PriceSparkline(
 
         fun norm(idx: Int): Pair<Float, Float> {
             val x = if (width <= 0f) 0f else idx / (prices.size - 1f) * width
-            val raw = (prices[idx] - minP) / range
+            val raw = ((prices[idx] - minP) / range).toFloat()
             // flip + add vertical breathing room
             val y = height - (raw * (height - 12f)) - 6f
             return x to y
